+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - Method `Realm.privilges()` to compute privileges on a Realm, a Realm object schema, or a Realm object. The method returns either a `Realm.Permissions.Realm` or `Realm.Permissions.Class` object.
   - For non-synced Realms, all privileges are always granted.
   - For more details, please read the reference documentation.
+* Added `Realm.default()` which will return the default synced Realm instance (#1688).
 
 ### Internal
 * Updated to Realm Core 5.3.0.